Allow choice of which lasio index to use
This adds a keyword argument "index" to well.from_lasio and
well.from_las. By default the value of this is None, which
replicates welly's existing behaviour.

The values this keyword argument can take are:

- "existing": welly simply uses the first column of the LAS file
  as the index. It could be in metres, feet, inches, light years,
  or seconds - no assumptions are made.
- "m": welly trusts that lasio can correctly identify the LAS file
  depth unit, and converts it to metres using lasio's "depth_m"
  attribute
- "ft": welly trusts that lasio can correctly identify the LAS file
  depth unit, and converts it to feet using lasio's "depth_ft"
  attribute

If index == None, it will be changed to "m", as this is welly's
current behaviour.