Refactor compute_position_log
In the process of moving welly to use wellpathpy to handle all
deviation and position data, the API has not been changed in
location.py, rather the compute_position_log function now uses
wellpathpy to return the same objects as it did previously.

There are no breaking changes but there are now additional methods
available to welly to compute a position log.

The same adjustments to TD, surface and azimuth_datum that were
made previously have simply been included in the new function and
remain in welly, rather than being added to wellpathpy.